+++ Kernel/Modules/AgentTicketHistory.pm
@@ -156,7 +156,7 @@ sub Run {
             # Make sure that the order of the values is correct, because we're now
             #   also showing the old ticket type on 'TypeUpdate'.
             elsif ( $Data->{HistoryType} eq 'TypeUpdate' ) {
-                @Values = ( $Values[2], $Values[3], $Values[0], $Values[1] );
+                @Values = ( $Values[2] // '', $Values[3] // '', $Values[0], $Values[1] );
             }
 
             $Data->{Name} = $LayoutObject->{LanguageObject}->Translate(